renamed EditDataButton to FormEditDescription and other bits in it fixed a bug in description update logic started moving api requests out of components into a separate service
